pyproject.toml
[tool.poetry]
name = "pdf-mail-merger"
version = "0.1.0"
description = ""
license = "AGPL-3.0-or-later"
authors = ["manhtai"]

[tool.poetry.dependencies]
python = "^3.9"
gunicorn = "^20.1.0"
django = "3.2"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
django-grappelli = "^2.14.3"
reportlab = "^3.5.66"
PyPDF2 = "^1.26.0"
arabic-reshaper = "^2.1.3"
python-bidi = "^0.4.2"
python-dotenv = "^0.17.0"
heroicons = {extras = ["django"], version = "^1.0.1"}
sentry-sdk = "^1.0.0"
django-q = "^1.3.5"
croniter = "^1.0.12"
django-storages = "^1.11.1"
boto3 = "^1.17.54"
django-simple-captcha = "^0.5.14"
tablib = {extras = ["xlsx", "xls"], version = "^3.0.0"}


[tool.poetry.dev-dependencies]
honcho = "^1.0.1"
ipython = "^7.22.0"
[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
